+++ src/hooks.js
@@ -11,6 +11,23 @@ export function init() {
   registerSettings();
 }
 
+function getDataEditField() {
+  let dataEditField;
+  switch (game.system.id) {
+    case "yzecoriolis":
+      dataEditField = "system.keyArt";
+      break;
+    default:
+      dataEditField = "img";
+  }
+  return dataEditField;
+}
+
+function getAvatarPath(actor) {
+  const key = getDataEditField();
+  return getProperty(actor, key);
+}
+
 /**
  * Launch the tokenizer
  * Options include
@@ -40,41 +57,25 @@ function launchTokenizer(options, callback) {
 
 async function updateActor(tokenizerResponse) {
   logger.debug("Updating Actor, tokenizer data", tokenizerResponse);
-
   const dateTag = `${+new Date()}`;
 
-  const version = (game.version ?? game.data.version);
-  const v10 = Utils.versionCompare(version, "10.0") >= 0;
-
-  const actorDataPrototypeToken = v10
-    ? tokenizerResponse.actor.prototypeToken
-    : tokenizerResponse.actor.data.token;
-
-  const originalPrototypeTokenPath = v10
-    ? actorDataPrototypeToken.texture.src
-    : actorDataPrototypeToken.img;
-
   // updating the avatar filename
-  const update = {
-    img: tokenizerResponse.avatarFilename.split("?")[0] + "?" + dateTag,
-  };
+  const update = {};
+  const avatarKey = getDataEditField(tokenizerResponse.actor);
+  update[avatarKey] = tokenizerResponse.avatarFilename.split("?")[0] + "?" + dateTag;
 
-  if (!actorDataPrototypeToken.randomImg) {
+  if (!tokenizerResponse.actor.prototypeToken.randomImg) {
     // for non-wildcard tokens, we set the token img now
     const tokenPath = tokenizerResponse.tokenFilename.split("?")[0] + "?" + dateTag;
-    if (v10) {
-      setProperty(update, "prototypeToken.texture.src", tokenPath);
-    } else {
-      setProperty(update, "token.img", tokenPath);
-    }
-  } else if (originalPrototypeTokenPath.indexOf("*") === -1) {
+    setProperty(update, "prototypeToken.texture.src", tokenPath);
+  } else if (tokenizerResponse.actor.prototypeToken.texture.src.indexOf("*") === -1) {
     // if it is a wildcard and it isn't get like one, we change that
     const actorName = tokenizerResponse.actor.name.replace(/[^\w.]/gi, "_").replace(/__+/g, "");
     const options = DirectoryPicker.parse(tokenizerResponse.tokenUploadDirectory);
 
     // set it to a wildcard we can actually use
     const imageFormat = game.settings.get(CONSTANTS.MODULE_ID, "image-save-type");
-    ui.notifications.info("Tokenizer: Wildcarding token image to " + originalPrototypeTokenPath);
+    ui.notifications.info("Tokenizer: Wildcarding token image to " + tokenizerResponse.actor.prototypeToken.texture.src);
     update.token = {
       img: `${options.current}/${actorName}.Token-*.${imageFormat}`,
     };
@@ -85,8 +86,7 @@ async function updateActor(tokenizerResponse) {
   await tokenizerResponse.actor.update(update);
   // if there is a scene token, lets update it
   if (tokenizerResponse.token) {
-    const tokenUpdate = v10 ? update.prototypeToken : update.token;
-    tokenizerResponse.token.update(tokenUpdate);
+    tokenizerResponse.token.update(update.prototypeToken);
   }
 }
 
@@ -95,17 +95,14 @@ function tokenizeActor(actor) {
     ui.notifications.warn(game.i18n.localize(`${CONSTANTS.MODULE_ID}.requires-upload-permission`));
   }
 
-  const version = (game.version ?? game.data.version);
-  const v10 = Utils.versionCompare(version, "10.0") >= 0;
-
   const options = {
     actor: actor,
     name: actor.name,
-    type: actor.data.type === "character" ? "pc" : "npc",
-    disposition: v10 ? actor.prototypeToken.disposition : actor.data.token.disposition,
-    avatarFilename: actor.img,
-    tokenFilename: v10 ? actor.prototypeToken.texture.src : actor.data.token.img,
-    isWildCard: v10 ? actor.prototypeToken.randomImg : actor.data.token.randomImg,
+    type: actor.type === "character" ? "pc" : "npc",
+    disposition: actor.prototypeToken.disposition,
+    avatarFilename: getAvatarPath(actor),
+    tokenFilename: actor.prototypeToken.texture.src,
+    isWildCard: actor.prototypeToken.randomImg,
   };
 
   launchTokenizer(options, updateActor);
@@ -117,17 +114,14 @@ function tokenizeSceneToken(doc) {
     ui.notifications.warn(game.i18n.localize(`${CONSTANTS.MODULE_ID}.requires-upload-permission`));
   }
 
-  const version = (game.version ?? game.data.version);
-  const v10 = Utils.versionCompare(version, "10.0") >= 0;
-
   const options = {
     actor: doc.actor,
     token: doc.token,
     name: doc.token.name,
-    type: doc.actor.data.type === "character" ? "pc" : "npc",
-    disposition: v10 ? doc.token.disposition : doc.token.data.disposition,
-    avatarFilename: v10 ? doc.actor.img : doc.actor.data.img,
-    tokenFilename: v10 ? doc.token.texture.src : doc.token.data.img,
+    type: doc.actor.type === "character" ? "pc" : "npc",
+    disposition: doc.token.disposition,
+    avatarFilename: getAvatarPath(doc.actor),
+    tokenFilename: doc.token.texture.src,
     nameSuffix: `${doc.token.id}`,
   };
 
@@ -144,41 +138,26 @@ function tokenizeDoc(doc) {
 }
 
 async function updateSceneTokenImg(actor) {
-  const version = (game.version ?? game.data.version);
-  const v10 = Utils.versionCompare(version, "10.0") >= 0;
-
   const updates = await Promise.all(actor.getActiveTokens().map(async (t) => {
     const newToken = await actor.getTokenData();
-    const tokenUpdate = v10
-      ? {
-        _id: t.id,
-        "texture.src": newToken.texture.src,
-      }
-      : {
-        _id: t.id,
-        img: newToken.img,
-      };
+    const tokenUpdate = {
+      _id: t.id,
+      "texture.src": newToken.texture.src,
+    };
     return tokenUpdate;
   }));
   if (updates.length) canvas.scene.updateEmbeddedDocuments("Token", updates);
 }
 
 export async function autoToken(actor, options) {
-  const version = (game.version ?? game.data.version);
-  const v10 = Utils.versionCompare(version, "10.0") >= 0;
- 
-  // construct our 
-  const actorData = v10
-   ? actor
-   : actor.data.data ? actor.data : actor;
   const defaultOptions = {
     actor: actor,
-    name: actorData.name,
-    type: actorData.type === "character" ? "pc" : "npc",
-    disposition: v10 ? actorData.prototypeToken.disposition : actorData.token.disposition,
-    avatarFilename: actorData.img,
-    tokenFilename: v10 ? actorData.prototypeToken.texture.src : actorData.token.img,
-    isWildCard: v10 ? actorData.prototypeToken.randomImg : actorData.token.randomImg,
+    name: actor.name,
+    type: actor.type === "character" ? "pc" : "npc",
+    disposition: actor.prototypeToken.disposition,
+    avatarFilename: getAvatarPath(actor),
+    tokenFilename: actor.prototypeToken.texture.src,
+    isWildCard: actor.prototypeToken.randomImg,
     auto: true,
     updateActor: true,
     // tokenOffset: { position: { x: -35, y: -35 } },
@@ -241,11 +220,7 @@ export function ready() {
   sheetNames.forEach((sheetName) => {
     Hooks.on("render" + sheetName, (app, html, data) => {
       if (game.user) {
-        const version = game.version ?? game.data.version;
-        const doc = isNewerVersion(version, "0.8.2") 
-        // is this token on a scene, if so we need to handle updates differently
-          ? (app.token) ? app : app.document
-          : app.entity;
+        const doc = (app.token) ? app : app.document;
 
         if (titleLink) {
           const button = $(`<a class="header-button ${CONSTANTS.MODULE_ID}" id="${CONSTANTS.MODULE_ID}-button" title="Tokenizer"><i class="far fa-user-circle"></i> Tokenizer</a>`);
@@ -265,8 +240,9 @@ export function ready() {
           ? disableAvatarClickGlobal
           : disableAvatarClickUser === "default";
 
+        const dataEditField = getDataEditField();
         $(html)
-        .find(`[data-edit=img]`)
+        .find(`[data-edit="${dataEditField}"]`)
         .each((index, element) => {
           // deactivating the original FilePicker click
           $(element).off("click");
@@ -286,14 +262,14 @@ export function ready() {
               // showing the filepicker
               new FilePicker({
                 type: "image",
-                current: data.actor.data.img,
+                current: data.actor[dataEditField],
                 callback: (path) => {
                   event.currentTarget.src = path;
                   app._onSubmit(event);
                 },
                 top: app.position.top + 40,
                 left: app.position.left + 10,
-              }).browse(data.actor.data.img);
+              }).browse(data.actor[dataEditField]);
             }
           });
         });